requirements.txt
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
gspread>=5.12.0
moviepy>=1.0.3
Pillow>=10.0.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
